refactor(openclaw): extract history reconciliation helpers
