+++ src/main/libs/agentEngine/openclawRuntimeAdapter.ts
@@ -77,6 +77,21 @@ import {
 import { buildCoworkTopKEvidenceBridgeResult } from './coworkTopKEvidence';
 import { buildCoworkWorkspaceRehydrationBridge } from './coworkWorkspaceRehydration';
 import { OpenClawApprovalController } from './openclawApprovalController';
+import {
+  applyLocalTimestampsToEntries,
+  buildGatewayMediaMetadata,
+  findTailAlignment,
+  getLocalMediaAttachmentsKey,
+  isSameReconciledEntry,
+  type ReconciledConversationEntry,
+} from './openclawConversationReconciliation';
+import {
+  buildCronRunHistoryEntries,
+  buildCronRunLocalHistoryEntries,
+  findCronRunHistoryLocalMatch,
+  hasCronRunHistoryForSession,
+  shouldReplaceLocalConversationWithCronHistory,
+} from './openclawCronRunHistorySync';
 import { SubagentTracker } from './subagentTracker';
 import type {
   CoworkContextUsage,
@@ -194,55 +209,6 @@ function validateRuntimeImageAttachments(
   }
   return { ok: true };
 }
-
-const normalizeLocalMediaPathKey = (value: unknown): string => {
-  if (typeof value !== 'string') return '';
-  return value.trim().replace(/\\/g, '/').toLowerCase();
-};
-
-const getLocalMediaAttachmentsKey = (metadata: unknown): string => {
-  if (!isRecord(metadata) || !Array.isArray(metadata.localMediaAttachments)) {
-    return '';
-  }
-  return metadata.localMediaAttachments
-    .map((item) => {
-      if (!isRecord(item)) return '';
-      const localPath = normalizeLocalMediaPathKey(item.localPath);
-      if (!localPath) return '';
-      const mimeType = typeof item.mimeType === 'string' ? item.mimeType.trim().toLowerCase() : '';
-      return `${localPath}\x1e${mimeType}`;
-    })
-    .filter(Boolean)
-    .sort()
-    .join('\x1f');
-};
-
-const buildGatewayMediaMetadata = (
-  entry: { mediaAttachments?: Array<{ localPath: string; mimeType?: string }> },
-): Record<string, unknown> | undefined => {
-  const attachments = entry.mediaAttachments
-    ?.map((attachment) => {
-      const localPath = attachment.localPath.trim();
-      if (!localPath) return null;
-      const mimeType = attachment.mimeType?.trim();
-      return {
-        localPath,
-        ...(mimeType ? { mimeType } : {}),
-        name: path.basename(localPath),
-      };
-    })
-    .filter((attachment): attachment is { localPath: string; mimeType?: string; name: string } => attachment !== null);
-
-  return attachments?.length ? { localMediaAttachments: attachments } : undefined;
-};
-
-const isSameReconciledEntry = (
-  left: { role: 'user' | 'assistant'; text: string; metadata?: Record<string, unknown> },
-  right: { role: 'user' | 'assistant'; text: string; metadata?: Record<string, unknown> },
-): boolean => {
-  return isSameHistoryEntry(left, right)
-    && getLocalMediaAttachmentsKey(left.metadata) === getLocalMediaAttachmentsKey(right.metadata);
-};
 
 /** How we chose assistant text to persist at chat.final (for tests and logs). */
 export type PersistedSegmentPickReason =
@@ -821,18 +787,6 @@ type ChannelHistorySyncEntry = {
   metadata?: Record<string, unknown>;
 };
 
-type ReconciledConversationEntry = {
-  role: 'user' | 'assistant';
-  text: string;
-  metadata?: Record<string, unknown>;
-  timestamp?: number;
-};
-
-const CronRunHistoryMetadataKey = {
-  SessionKey: 'openclawCronRunSessionKey',
-  EntryIndex: 'openclawCronRunEntryIndex',
-} as const;
-
 const isRecord = (value: unknown): value is Record<string, unknown> => {
   return Boolean(value && typeof value === 'object' && !Array.isArray(value));
 };
@@ -998,174 +952,6 @@ const normalizeEntryText = (
   if (flags.isPopo && role === 'user') result = stripPopoSystemHeader(result);
   if (flags.isFeishu && role === 'user') result = stripFeishuSystemHeader(result);
   return result;
-};
-
-const isSameHistoryEntry = (
-  left: { role: 'user' | 'assistant'; text: string },
-  right: { role: 'user' | 'assistant'; text: string },
-): boolean => left.role === right.role && left.text === right.text;
-
-const historyEntryKey = (entry: { role: 'user' | 'assistant'; text: string }): string => {
-  return `${entry.role}\x1f${entry.text}`;
-};
-
-const isValidMessageTimestamp = (value: unknown): value is number => {
-  return typeof value === 'number' && Number.isFinite(value) && value > 0;
-};
-
-const applyLocalTimestampsToEntries = (
-  entries: ReconciledConversationEntry[],
-  localEntries: Array<{ role: 'user' | 'assistant'; text: string; timestamp?: number }>,
-): ReconciledConversationEntry[] => {
-  const localTimestamps = new Map<string, number[]>();
-  for (const entry of localEntries) {
-    if (!isValidMessageTimestamp(entry.timestamp)) continue;
-    const key = historyEntryKey(entry);
-    const timestamps = localTimestamps.get(key) ?? [];
-    timestamps.push(entry.timestamp);
-    localTimestamps.set(key, timestamps);
-  }
-
-  return entries.map((entry) => {
-    if (isValidMessageTimestamp(entry.timestamp)) {
-      return entry;
-    }
-    const timestamps = localTimestamps.get(historyEntryKey(entry));
-    const timestamp = timestamps?.shift();
-    return timestamp != null ? { ...entry, timestamp } : entry;
-  });
-};
-
-const getCronRunHistorySessionKey = (metadata: unknown): string | null => {
-  if (!isRecord(metadata)) return null;
-  const value = metadata[CronRunHistoryMetadataKey.SessionKey];
-  return typeof value === 'string' && value.trim() ? value.trim() : null;
-};
-
-const withCronRunHistoryMetadata = (
-  entry: ReconciledConversationEntry,
-  sessionKey: string,
-  entryIndex: number,
-): ReconciledConversationEntry => ({
-  ...entry,
-  metadata: {
-    ...(entry.metadata ?? {}),
-    [CronRunHistoryMetadataKey.SessionKey]: sessionKey,
-    [CronRunHistoryMetadataKey.EntryIndex]: entryIndex,
-  },
-});
-
-const isLocalConversationCoveredByCronHistory = (
-  localEntries: ReadonlyArray<{ role: 'user' | 'assistant'; text: string; metadata?: Record<string, unknown> }>,
-  authoritativeEntries: ReadonlyArray<ReconciledConversationEntry>,
-): boolean => {
-  if (localEntries.length > authoritativeEntries.length) return false;
-
-  let authIdx = 0;
-  for (const local of localEntries) {
-    let matched = false;
-    while (authIdx < authoritativeEntries.length) {
-      const authoritative = authoritativeEntries[authIdx++];
-      if (isSameHistoryEntry(local, authoritative)) {
-        matched = true;
-        break;
-      }
-    }
-    if (!matched) return false;
-  }
-  return true;
-};
-
-/**
- * Find the tail-alignment point between local and authoritative entries.
- *
- * `chat.history` can return a bounded tail window that starts in the middle of
- * a turn, often with an assistant entry before the first user anchor. Prefer a
- * full role/text overlap first; then fall back to user-message anchors and
- * report both the local and authoritative start indices so leading orphan
- * assistant entries are not duplicated into the local prefix on every poll.
- */
-const findTailAlignment = (
-  localEntries: ReadonlyArray<{ role: 'user' | 'assistant'; text: string }>,
-  authEntries: ReadonlyArray<{ role: 'user' | 'assistant'; text: string }>,
-): { localIdx: number; authIdx: number } | null => {
-  if (authEntries.length === 0) return null;
-  if (localEntries.length === 0) return { localIdx: 0, authIdx: 0 };
-
-  const maxEntryOverlap = Math.min(localEntries.length, authEntries.length);
-  for (let overlap = maxEntryOverlap; overlap >= 1; overlap -= 1) {
-    const localStart = localEntries.length - overlap;
-    let match = true;
-    for (let idx = 0; idx < overlap; idx += 1) {
-      if (!isSameHistoryEntry(localEntries[localStart + idx], authEntries[idx])) {
-        match = false;
-        break;
-      }
-    }
-    if (match) {
-      return { localIdx: localStart, authIdx: 0 };
-    }
-  }
-
-  // Extract user-only entries with their original indices
-  const localUsers: Array<{ idx: number; text: string }> = [];
-  for (let i = 0; i < localEntries.length; i++) {
-    if (localEntries[i].role === 'user') {
-      localUsers.push({ idx: i, text: localEntries[i].text });
-    }
-  }
-
-  const authUsers: Array<{ idx: number; text: string }> = [];
-  for (let i = 0; i < authEntries.length; i++) {
-    const entry = authEntries[i];
-    if (entry.role === 'user') {
-      authUsers.push({ idx: i, text: entry.text });
-    }
-  }
-
-  if (authUsers.length === 0 || localUsers.length === 0) {
-    // No user messages to anchor on — fall back to full replace
-    return { localIdx: 0, authIdx: 0 };
-  }
-
-  // Find largest k where localUsers tail-k texts == authUsers head-k texts
-  const maxK = Math.min(localUsers.length, authUsers.length);
-  for (let k = maxK; k >= 1; k--) {
-    const localStart = localUsers.length - k;
-    let match = true;
-    for (let j = 0; j < k; j++) {
-      if (localUsers[localStart + j].text !== authUsers[j].text) {
-        match = false;
-        break;
-      }
-    }
-    if (match) {
-      // The alignment point is the original index of the first overlapping
-      // user message in local and authoritative history.
-      const localIdx = localUsers[localStart].idx;
-      const authIdx = authUsers[0].idx;
-      if (authIdx > 0) {
-        const leadingLocalIdx = localIdx - authIdx;
-        const leadingAuthAlreadyPresent = leadingLocalIdx >= 0
-          && authEntries.slice(0, authIdx).every((entry, idx) =>
-            isSameHistoryEntry(localEntries[leadingLocalIdx + idx], entry),
-          );
-        if (!leadingAuthAlreadyPresent) {
-          return {
-            localIdx: Math.max(0, leadingLocalIdx),
-            authIdx: 0,
-          };
-        }
-      }
-      return {
-        localIdx,
-        authIdx,
-      };
-    }
-  }
-
-  // No overlap found
-  return null;
 };
 
 const extractMessageText = extractGatewayMessageText;
@@ -7779,35 +7565,7 @@ export class OpenClawRuntimeAdapter extends EventEmitter implements CoworkRuntim
         previousCount: previousHistoryCount,
       });
 
-      const authoritativeEntries: ReconciledConversationEntry[] = [];
-      for (const entry of extractGatewayHistoryEntries(history.messages)) {
-        const role = entry.role;
-        if (role !== 'user' && role !== 'assistant') continue;
-        const text = entry.text.trim();
-        if (!text || shouldSuppressHeartbeatText(role, text)) continue;
-
-        let metadata: Record<string, unknown> | undefined;
-        if (role === 'assistant' && (entry.usage || entry.model)) {
-          metadata = {};
-          if (entry.usage) {
-            metadata.usage = {
-              ...(entry.usage.input != null && { inputTokens: entry.usage.input }),
-              ...(entry.usage.output != null && { outputTokens: entry.usage.output }),
-            };
-          }
-          if (entry.model) {
-            metadata.model = entry.model;
-          }
-        }
-
-        authoritativeEntries.push(withCronRunHistoryMetadata({
-          role,
-          text,
-          ...(metadata && { metadata }),
-          ...(entry.timestamp != null && { timestamp: entry.timestamp }),
-        }, sessionKey, authoritativeEntries.length));
-      }
-
+      const authoritativeEntries = buildCronRunHistoryEntries(history.messages, sessionKey);
       if (authoritativeEntries.length === 0) {
         console.log('[CronHistorySync] no user/assistant entries in history - sessionId:', sessionId);
         this.channelSyncCursor.set(sessionId, 0);
@@ -7817,33 +7575,13 @@ export class OpenClawRuntimeAdapter extends EventEmitter implements CoworkRuntim
       const session = this.store.getSession(sessionId);
       if (!session) return;
 
-      if (session.messages.some((message: CoworkMessage) =>
-        getCronRunHistorySessionKey(message.metadata) === sessionKey,
-      )) {
+      if (hasCronRunHistoryForSession(session.messages, sessionKey)) {
         this.channelSyncCursor.set(sessionId, authoritativeEntries.length);
         return;
       }
 
-      const localEntries = session.messages
-        .filter((message: CoworkMessage) => message.type === 'user' || message.type === 'assistant')
-        .map((message: CoworkMessage) => ({
-          id: message.id,
-          role: message.type as 'user' | 'assistant',
-          text: message.content.trim(),
-          metadata: isRecord(message.metadata) ? message.metadata : undefined,
-          timestamp: message.timestamp,
-        }))
-        .filter((entry) => entry.text && !shouldSuppressHeartbeatText(entry.role, entry.text));
-
-      const hasOtherCronRunHistory = localEntries.some((entry) => {
-        const importedSessionKey = getCronRunHistorySessionKey(entry.metadata);
-        return Boolean(importedSessionKey && importedSessionKey !== sessionKey);
-      });
-
-      if (
-        !hasOtherCronRunHistory
-        && isLocalConversationCoveredByCronHistory(localEntries, authoritativeEntries)
-      ) {
+      const localEntries = buildCronRunLocalHistoryEntries(session.messages);
+      if (shouldReplaceLocalConversationWithCronHistory(localEntries, authoritativeEntries, sessionKey)) {
         this.store.replaceConversationMessages(
           sessionId,
           applyLocalTimestampsToEntries(authoritativeEntries, localEntries),
@@ -7854,12 +7592,12 @@ export class OpenClawRuntimeAdapter extends EventEmitter implements CoworkRuntim
 
       const usedLocalMessageIds = new Set<string>();
       for (const authoritative of authoritativeEntries) {
-        const matchingLocal = localEntries.find((entry) => {
-          if (usedLocalMessageIds.has(entry.id)) return false;
-          if (!isSameHistoryEntry(entry, authoritative)) return false;
-          const importedSessionKey = getCronRunHistorySessionKey(entry.metadata);
-          return !importedSessionKey || importedSessionKey === sessionKey;
-        });
+        const matchingLocal = findCronRunHistoryLocalMatch(
+          authoritative,
+          localEntries,
+          usedLocalMessageIds,
+          sessionKey,
+        );
 
         if (matchingLocal) {
           usedLocalMessageIds.add(matchingLocal.id);
